+++ frontend/common/services/useOrganisationUsage.ts
@@ -28,7 +28,6 @@ export const organisationUsageService = service
           let environmentDocument = 0
           let identities = 0
           data?.map((v) => {
-            environmentDocument += v.environment_document || 0
             flags += v.flags || 0
             traits += v.traits || 0
             environmentDocument += v.environment_document || 0
@@ -57,14 +56,11 @@ export async function getOrganisationUsage(
     typeof organisationUsageService.endpoints.getOrganisationUsage.initiate
   >[1],
 ) {
-  store.dispatch(
+  return store.dispatch(
     organisationUsageService.endpoints.getOrganisationUsage.initiate(
       data,
       options,
     ),
-  )
-  return Promise.all(
-    store.dispatch(organisationUsageService.util.getRunningQueriesThunk()),
   )
 }
 // END OF FUNCTION_EXPORTS